# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Tech-Ops by Martin PMO API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /cases:
    get:
      operationId: listCases
      tags: [cases]
      summary: List user diagnostic cases
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [open, in_progress, resolved, closed]
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of cases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DiagnosticCase"
    post:
      operationId: createCase
      tags: [cases]
      summary: Create a new diagnostic case
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCaseBody"
      responses:
        "201":
          description: Created case
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiagnosticCase"
  /cases/{id}:
    get:
      operationId: getCase
      tags: [cases]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Case details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiagnosticCase"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    patch:
      operationId: updateCase
      tags: [cases]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCaseBody"
      responses:
        "200":
          description: Updated case
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiagnosticCase"
  /cases/{id}/diagnose:
    post:
      operationId: runDiagnostic
      tags: [cases]
      summary: Run Apphia diagnostic pipeline on a case
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: SSE stream of diagnostic progress
          content:
            text/event-stream: {}
  /cases/batch:
    post:
      operationId: runBatchDiagnostics
      tags: [cases]
      summary: Run parallel batch diagnostics on multiple cases
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchDiagnosticsBody"
      responses:
        "200":
          description: SSE stream of batch progress
          content:
            text/event-stream: {}
...
  /dashboard/activity:
    get:
      operationId: getRecentActivity
      tags: [dashboard]
      summary: Get recent activity feed
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Recent activity
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"
...
  /automation/rules:
    get:
      operationId: listAutomationRules
      tags: [automation]
      summary: List all automation rules
      responses:
        "200":
          description: List of rules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AutomationRuleResponse"
    post:
      operationId: createAutomationRule
      tags: [automation]
      summary: Create an automation rule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAutomationRuleBody"
      responses:
        "201":
          description: Created rule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AutomationRuleResponse"
  /automation/rules/{id}:
    patch:
      operationId: updateAutomationRule
      tags: [automation]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAutomationRuleBody"
      responses:
        "200":
          description: Updated rule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AutomationRuleResponse"
    delete:
      operationId: deleteAutomationRule
      tags: [automation]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    DiagnosticCase:
      type: object
      required: [id, userId, title, status, createdAt, updatedAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        title:
          type: string
        description:
          type: [string, 'null']
        status:
          type: string
          enum: [open, in_progress, resolved, closed]
        priority:
          type: [string, 'null']
        diagnosticTier:
          type: [integer, 'null']
        rootCause:
          type: [string, 'null']
        resolution:
          type: [string, 'null']
        signals:
          type: ['object', 'null']
        udoPath:
          type: ['object', 'null']
        confidenceScore:
          type: [integer, 'null']
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        resolvedAt:
          type: [string, 'null']
          format: date-time
    CreateCaseBody:
      type: object
      required: [title]
      properties:
        title:
          type: string
        description:
          type: string
        priority:
          type: string
          enum: [low, medium, high, critical]
    UpdateCaseBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [open, in_progress, resolved, closed]
        priority:
          type: string
          enum: [low, medium, high, critical]
        resolution:
          type: string
    BatchDiagnosticsBody:
      type: object
      required: [caseIds]
      properties:
        caseIds:
          type: array
          items:
            type: integer
...
    ActivityItem:
      type: object
      required: [id, type, message, timestamp]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [case_created, case_resolved, diagnostic_run, automation_triggered, connector_alert]
        message:
          type: string
        timestamp:
          type: string
          format: date-time
        caseId:
          type: [integer, 'null']
...
    AutomationRuleResponse:
      type: object
      required: [id, userId, name, trigger, action, enabled, createdAt, updatedAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        name:
          type: string
        description:
          type: [string, 'null']
        trigger:
          type: string
        action:
          type: string
        enabled:
          type: string
        permissions:
          type: ['object', 'null']
        executionCount:
          type: [integer, 'null']
        lastExecuted:
          type: [string, 'null']
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateAutomationRuleBody:
      type: object
      required: [name, trigger, action]
      properties:
        name:
          type: string
        description:
          type: string
        trigger:
          type: string
        action:
          type: string
        permissions:
          type: object
    UpdateAutomationRuleBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        trigger:
          type: string
        action:
          type: string
        enabled:
          type: string
        permissions:
          type: object
...